+++ tests/00.complete-path.test.js
@@ -1,4 +1,5 @@
 const sw = require('starkware_crypto')
+const getTokenAddressFromTokenInfoOrThrow = require('../src/lib/dvf/token/getTokenAddressFromTokenInfoOrThrow')
 const { setup, register, deposit, withdraw, drip, testOrders, sleep } = require('./helpers')
 
 describe.skip('00 - Complete Path', () => {
@@ -36,7 +37,8 @@ describe.skip('00 - Complete Path', () => {
     const token = 'USDT'
 
     // get some tokens from the faucet
-    await drip(dvf, dvf.config.tokenRegistry[token].tokenAddress, address)
+    const tokenAddress = getTokenAddressFromTokenInfoOrThrow(dvf.config.tokenRegistry[token], 'ETHEREUM')
+    await drip(dvf, tokenAddress, address)
 
     await dvf.contract.approve(token)
     const tokenDepositResponse = await deposit(dvf, token, 100, privateKey)