+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.4.0 - 2021-06-4
+### Changed
+- Accomodating changes in configuration tokenRegistry structure (`tokenAddressPerChain`)
+- `dvf.approve` to accept an extra argument `chain` (defaults to `'ETHEREUM'`)
+
 ## 2.3.0 - 2021-05-25
 ### Changed
 - `dvf.depositV2` and `dvf.registerAndDeposit` to accept optional `web3Options` parameter for internal use (better gas limit estimates)